requirements.txt
folium==0.17.0
haversine==2.8.1
nbformat==5.10.4
numpy==2.1.1
pandas==2.2.2
Pillow==10.4.0
plotly==5.24.0
streamlit==1.38.0
streamlit_folium==0.22.0
